More cleanup of old source; move compat check to differences class
